instruction: Derive everything for wasm build (#3606)
#### Problem

As noticed in
https://github.com/solana-labs/solana-program-library/pull/7489, the
wasm build for `Instruction` doesn't contain `Clone`. This is because we
missed the derivations for the wasm version of `Instruction` in
https://github.com/anza-xyz/agave/pull/1658.

#### Summary of changes

Looking back through that PR, I noticed that we only missed those
additional derives on `Instruction`, so just fix those in `Instruction`.

Since this is fixing a small regression for wasm builds, I think we
should backport the change to v2.1.

(cherry picked from commit d3bce9f02414e47b7f498d71503f8061a75d72b0)